fixing node bugs

+++ StishBoard/StishBoard/StishMiniMaxNode.cs
@@ -38,34 +38,34 @@ namespace StishBoard
             Allegiance = PlayersTurn;
         }
 
-        private int P1BarracksHealth;
-        private int P1UnitHealth;
-        private int P1BaseHealth;
-        private int P1BarracksNumber;
-        private int P1Balance;
+        private uint P1BarracksHealth;
+        private uint P1UnitHealth;
+        private uint P1BaseHealth;
+        private uint P1BarracksNumber;
+        private uint P1Balance;
 
-        private int P2BarracksHealth;
-        private int P2UnitHealth;
-        private int P2BaseHealth;
-        private int P2BarracksNumber;
-        private int P2Balance;       
+        private uint P2BarracksHealth;
+        private uint P2UnitHealth;
+        private uint P2BaseHealth;
+        private uint P2BarracksNumber;
+        private uint P2Balance;       
 
         public void FindValue(TreeNode Parent, BoardState PassedBoardState)
         {
             //parent value, current barracks of both players, base health of both players, health of units of both players
-            P1BarracksNumber = (int)PassedBoardState.Counting("Barracks", PassedBoardState.Player1, true);
-            P1BarracksHealth = (int)PassedBoardState.Counting("Barracks", PassedBoardState.Player1, false);
-            P1UnitHealth = (int)PassedBoardState.Counting("Unit", PassedBoardState.Player1, false);
-            P1BaseHealth = (int)PassedBoardState.Counting("Base", PassedBoardState.Player1, false);           
-            P1Balance = (int)PassedBoardState.Player1.Balance;
+            P1BarracksNumber = PassedBoardState.Counting("Barracks", PassedBoardState.Player1, true);
+            P1BarracksHealth = PassedBoardState.Counting("Barracks", PassedBoardState.Player1, false);
+            P1UnitHealth = PassedBoardState.Counting("Unit", PassedBoardState.Player1, false);
+            P1BaseHealth = PassedBoardState.Counting("Base", PassedBoardState.Player1, false);           
+            P1Balance = PassedBoardState.Player1.Balance;
 
-            P2BarracksNumber = (int)PassedBoardState.Counting("Barracks", PassedBoardState.Player2, true);
-            P2BaseHealth = (int)PassedBoardState.Counting("Base", PassedBoardState.Player2, false);
-            P2BarracksHealth = (int)PassedBoardState.Counting("Barracks", PassedBoardState.Player2, false);
-            P2UnitHealth = (int)PassedBoardState.Counting("Unit", PassedBoardState.Player2, false);           
-            P2Balance = (int)PassedBoardState.Player2.Balance;
+            P2BarracksNumber = PassedBoardState.Counting("Barracks", PassedBoardState.Player2, true);
+            P2BaseHealth = PassedBoardState.Counting("Base", PassedBoardState.Player2, false);
+            P2BarracksHealth = PassedBoardState.Counting("Barracks", PassedBoardState.Player2, false);
+            P2UnitHealth = PassedBoardState.Counting("Unit", PassedBoardState.Player2, false);           
+            P2Balance = PassedBoardState.Player2.Balance;
 
-            Value = (1 * (P1BarracksNumber - P2BarracksNumber)) + (1 * (P1BarracksHealth - P2BarracksHealth)) + (1 * (P1UnitHealth - P2UnitHealth)) + (1 * (P1BaseHealth - P2BaseHealth)) + (1 * (P1Balance - P2Balance));
+            Value = (1 * ((int)P1BarracksNumber - (int)P2BarracksNumber)) + (1 * ((int)P1BarracksHealth - (int)P2BarracksHealth)) + (1 * ((int)P1UnitHealth - (int)P2UnitHealth)) + (1 * ((int)P1BaseHealth - (int)P2BaseHealth)) + (1 * ((int)P1Balance - (int)P2Balance));
         }
 
         StishMiniMaxNode(TreeNode Parent, BoardState PassedBoardState) : base(Parent)
